backend/requirements.txt
anthropic>=0.40.0
openai>=1.30.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.7.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
chromadb>=0.5.0
feedparser>=6.0.0
